feat: implement middleware structure and routing for AI stages, including domain routing, headers, and preview handling
