+++ package.json
@@ -29,6 +29,7 @@
     "prep:download-translations": "jiti scripts/download-translations.ts",
     "prep:icons": "svg-sprite --symbol --symbol-dest='src/assets' --symbol-sprite=icon-sprite.svg src/assets/icons/*.svg && jiti scripts/icon-types.ts",
     "preview": "wrangler pages dev ./dist",
+    "start:ai-stages": "AI_STAGES_DEV=true astro dev",
     "lint": "run-s lint:* --print-label",
     "lint:astro": "astro check",
     "lint:eslint": "eslint . --ext .js,.ts,.astro",
